Separate table in NodeSearchPage to NodeTable
