+++ gui/gui/src/ts/models/Node.ts
@@ -1,0 +1,3 @@
+import {NodeFieldsFragment} from "api/queries/types/NodeFieldsFragment";
+
+export type Node = NodeFieldsFragment;
